requirements.txt
streamlit
chromadb
sentence-transformers
openai
llama-index-vector-stores-chroma
torch
